refact(*): Dockerfile and Helm chart updates
Reduced the number of layers at the Dockerfile by reestructuring the
COPY config for each file into one that is already at the expected
dir structure;

Dockerfile COPY commands for scripts were failing as it was using an
absolute path to /script-name.sh, which was replaced by its relative
equivalent;

Helm Chart updated to better support volumes. It can enable and disable
volumes on-demand. Previous to these changes there was no
persistence.enable=false because it would crash the running pod due to
invalid references to volumes.

volumeClaimTemplate added as a value to the Helm Chart to give
flexibility over the previously hard-coded namespace: infrastructure as
well as to give the freedom to disable this template and use regular
volumes.